requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asttokens==2.0.5
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
backcall==0.2.0
blis==0.7.7
catalogue==2.0.7
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
cymem==2.0.6
datasets==2.2.2
decorator==5.1.1
defusedxml==0.6.0
dill==0.3.4
docker-pycreds==0.4.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.3.0/en_core_web_sm-3.3.0-py3-none-any.whl
et-xmlfile==1.1.0
executing==0.8.3
filelock==3.7.0
frozenlist==1.3.0
fsspec==2022.5.0
gitdb==4.0.9
GitPython==3.1.27
huggingface-hub==0.6.0
idna==3.3
importlib-metadata==4.11.3
ipdb==0.13.9
ipykernel==5.3.4
ipython==7.33.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jupyter-client==6.1.7
jupyter-core==4.6.3
langcodes==3.3.0
lxml==4.5.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
multidict==6.0.2
multiprocess==0.70.12.2
murmurhash==1.0.7
nltk==3.5
numpy==1.21.6
openai==0.27.0
openpyxl==3.0.10
packaging==21.3
pandas==1.3.5
pandas-stubs==1.2.0.58
parso==0.8.3
pathtools==0.1.2
pathy==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
preshed==3.0.6
promise==2.3
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==8.0.0
pydantic==1.8.2
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-docx==0.8.10
pytz==2022.1
PyYAML==6.0
pyzmq==19.0.2
regex==2022.4.24
requests==2.27.1
responses==0.18.0
scikit-learn==1.0.2
scipy==1.7.3
sentry-sdk==1.5.12
seqeval==1.2.2
setproctitle==1.2.3
shortuuid==1.0.9
six==1.16.0
sklearn==0.0
smart-open==5.2.1
smmap==5.0.0
spacy==3.3.0
spacy-legacy==3.0.9
spacy-loggers==1.0.2
srsly==2.4.3
stack-data==0.2.0
thinc==8.0.16
threadpoolctl==3.1.0
tokenizers==0.12.1
toml==0.10.2
torch==1.11.0
tornado==6.0.4
tqdm==4.64.0
traitlets==5.2.1.post0
transformers==4.19.2
typer==0.4.1
typing-extensions==3.10.0.2
unrar==0.4
urllib3==1.26.9
wandb==0.12.16
wasabi==0.9.1
wcwidth==0.2.5
xxhash==3.0.0
yarl==1.7.2
zipp==2.2.0